Added done w
Main Menu and pause menu Yippie
