+++ Assets/Scenes/Main Menu.unity
@@ -326,6 +326,7 @@ GameObject:
   - component: {fileID: 928793293}
   - component: {fileID: 928793292}
   - component: {fileID: 928793291}
+  - component: {fileID: 928793294}
   m_Layer: 5
   m_Name: Start
   m_TagString: Untagged
@@ -396,7 +397,19 @@ MonoBehaviour:
   m_TargetGraphic: {fileID: 928793292}
   m_OnClick:
     m_PersistentCalls:
-      m_Calls: []
+      m_Calls:
+      - m_Target: {fileID: 928793294}
+        m_TargetAssemblyTypeName: MainMenu, Assembly-CSharp
+        m_MethodName: GoToGame
+        m_Mode: 1
+        m_Arguments:
+          m_ObjectArgument: {fileID: 0}
+          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
+          m_IntArgument: 0
+          m_FloatArgument: 0
+          m_StringArgument: 
+          m_BoolArgument: 0
+        m_CallState: 2
 --- !u!114 &928793292
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -435,6 +448,18 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 928793289}
   m_CullTransparentMesh: 1
+--- !u!114 &928793294
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 928793289}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ae861871608948346b82075211bb2221, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &1366179143
 GameObject:
   m_ObjectHideFlags: 0
@@ -919,6 +944,7 @@ GameObject:
   - component: {fileID: 2089986039}
   - component: {fileID: 2089986038}
   - component: {fileID: 2089986037}
+  - component: {fileID: 2089986040}
   m_Layer: 5
   m_Name: Quit
   m_TagString: Untagged
@@ -989,7 +1015,19 @@ MonoBehaviour:
   m_TargetGraphic: {fileID: 2089986038}
   m_OnClick:
     m_PersistentCalls:
-      m_Calls: []
+      m_Calls:
+      - m_Target: {fileID: 2089986040}
+        m_TargetAssemblyTypeName: MainMenu, Assembly-CSharp
+        m_MethodName: QuitGame
+        m_Mode: 1
+        m_Arguments:
+          m_ObjectArgument: {fileID: 0}
+          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
+          m_IntArgument: 0
+          m_FloatArgument: 0
+          m_StringArgument: 
+          m_BoolArgument: 0
+        m_CallState: 2
 --- !u!114 &2089986038
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1028,6 +1066,18 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 2089986035}
   m_CullTransparentMesh: 1
+--- !u!114 &2089986040
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2089986035}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ae861871608948346b82075211bb2221, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1660057539 &9223372036854775807
 SceneRoots:
   m_ObjectHideFlags: 0
